+++ RELEASENOTES.md
@@ -5,6 +5,9 @@
 * Text: Catch-and-log all fatal exceptions in `TextRenderer` instead of
   re-throwing, allowing playback to continue even if subtitles fail
   ([#6885](https://github.com/google/ExoPlayer/issues/6885)).
+* UI: Add an option to set whether to use the orientation sensor for rotation
+  in spherical playbacks
+  ([#6761](https://github.com/google/ExoPlayer/issues/6761)).
 * DASH: 
     * Update the manifest URI to avoid repeated HTTP redirects
       ([#6907](https://github.com/google/ExoPlayer/issues/6907)).